requirements.txt
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
pillow==10.3.0
tqdm~=4.66.4
httpx~=0.27.0
opencv-python-headless~=4.10.0.82
numpy~=1.26.4
insightface~=0.7.3
onnx~=1.16.0
onnxruntime~=1.18.0
requests~=2.32.2
python-dotenv~=1.0.1
torchmetrics~=1.4.0.post0
streamlit~=1.35.0
